+++ weldx/tests/transformations/test_local_cs.py
@@ -23,7 +23,6 @@ from weldx.transformations import WXRotation
 
 from ._util import check_coordinate_system, check_cs_close, r_mat_y, r_mat_z
 
-
 # test_init_time_formats ---------------------------------------------------------------
 
 timestamp = TS("2000-01-01")